// src/network/api/model/alarmDetailResponse.ts
/**
 * Generated by orval v6.18.1 🍺
 * Do not edit manually.
 * Springdoc
 * Springdoc을 사용한 Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { AlarmLocationDetailResponse } from "./alarmLocationDetailResponse";
import type { AlarmWeekResponse } from "./alarmWeekResponse";

export interface AlarmDetailResponse {
  /** 유저ID */
  userId?: number;
  /** 알람ID */
  alarmId?: number;
  /** 활성화여부(0 : 비활성화, 1 : 활성화) */
  enabled?: number;
  /** 진동여부(0: 비활성화, 1 : 활성화) */
  vibration?: number;
  /** 볼륨 */
  volume?: number;
  /** 다시울림 */
  reminder?: string;
  /** 알람시간(AM PM 시간) */
  alarmTime?: string;
  /** AM, PM */
  timeOfDay?: string;
  /** 알람지역목록 */
  alarmLocation?: AlarmLocationDetailResponse[];
  /** 알람요일목록 */
  alarmWeek?: AlarmWeekResponse[];
  /** 알람소리코드 */
  alarmSoundId?: number;
}